fix: JDA v5.5.0
Technically it was already supported, you were just met with an ugly
error due to the voice state rework which removed the #connectedMembers
field in audio channels.

All we did here was disable the problematic optimization (which is no
longer necessary) when running JDA 5.5.0 or greater.

The new field GuildImpl#voiceStateCache which was added to
replace #connectedMembers was automatically optimized so no additional
work necessary here